requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.3.1
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
billiard==4.2.1
binaryornot==0.4.4
Brotli==1.1.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cookiecutter==2.6.0
cryptography==43.0.1
cssselect2==0.7.0
decorator==5.1.1
Django==5.1
django-appconf==1.0.6
django-browser-reload==1.15.0
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
fonttools==4.53.1
frozenlist==1.4.1
future==1.0.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
html5lib==1.1
httplib2==0.22.0
idna==3.8
Jinja2==3.1.4
kombu==5.4.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
oauthlib==3.2.2
oscrypto==1.3.0
pillow==10.4.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
Pygments==2.18.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pyparsing==3.1.4
pypdf==5.0.1
pyphen==0.16.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
ratelim==0.1.6
rcssmin==1.1.2
redis==5.2.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.0
rjsmin==1.2.2
rsa==4.9
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.3.0
twilio==9.3.2
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
xhtml2pdf==0.2.16
XlsxWriter==3.2.0
yarl==1.13.1
zopfli==0.2.3
